api/requirements.txt
supabase==2.3.0
rdkit==2023.9.5
python-dotenv==1.0.0